+++ packages/react-pacer/src/async-queuer/useAsyncQueuer.ts
@@ -248,7 +248,6 @@ export function useAsyncQueuer<TValue, TSelected = {}>(
       mergedOptions,
     ) as unknown as ReactAsyncQueuer<TValue, TSelected>
 
-    /* eslint-disable-next-line @eslint-react/component-hook-factories -- Subscribe attached once in useState lazy init; stable per instance */
     asyncQueuerInstance.Subscribe = function Subscribe<TSelected>(props: {
       selector: (state: AsyncQueuerState<TValue>) => TSelected
       children: ((state: TSelected) => ReactNode) | ReactNode