+++ packages/react-pacer/src/async-throttler/useAsyncThrottler.ts
@@ -238,7 +238,6 @@ export function useAsyncThrottler<TFn extends AnyAsyncFunction, TSelected = {}>(
       mergedOptions,
     ) as unknown as ReactAsyncThrottler<TFn, TSelected>
 
-    /* eslint-disable-next-line @eslint-react/component-hook-factories -- Subscribe attached once in useState lazy init; stable per instance */
     asyncThrottlerInstance.Subscribe = function Subscribe<TSelected>(props: {
       selector: (state: AsyncThrottlerState<TFn>) => TSelected
       children: ((state: TSelected) => ReactNode) | ReactNode